+++ examples/copyContentToProjectFolder/applications/example/app/examples/events/connect.sse/get.js
@@ -1,4 +1,6 @@
 module.exports = function(client, callback) {
+
 	client.sse.channel = 'TestEvent';
 	callback();
+
 }