+++ examples/copyContentToProjectFolder/applications/example/app/examples/events/connect.ws/get.js
@@ -13,4 +13,5 @@ module.exports = function(client, callback) {
 	});
 
 	callback();
+
 }