+++ examples/copyContentToProjectFolder/applications/example/app/examples/mongodb/getCollections.json/get.js
@@ -1,4 +1,5 @@
 module.exports = function(client, callback) {
+
 	dbImpress.connection.collections(function(err, collections) {
 		var items = [];
 		for (var i = 0; i < collections.length; i++) {
@@ -7,4 +8,5 @@ module.exports = function(client, callback) {
 		client.context.data = items;
 		callback();
 	});
+
 }