+++ examples/copyContentToProjectFolder/applications/example/app/examples/mongodb/getData.json/get.js
@@ -1,6 +1,6 @@
 module.exports = function(client, callback) {
 
-	impress.security.db.sessions.find({}).toArray(function(err, nodes) {
+	dbAlias.testCollection.find({}).toArray(function(err, nodes) {
 		client.context.data = nodes;
 		callback();
 	});