+++ examples/copyContentToProjectFolder/applications/example/app/examples/mongodb/insertData.json/get.js
@@ -1,6 +1,8 @@
 module.exports = function(client, callback) {
+
 	dbAlias.testCollection.insert(client.query, function(err) {
 		client.context.data = !err;
 		callback();
 	});
+
 }