+++ examples/copyContentToProjectFolder/applications/example/app/examples/mysql/getCities.json/get.js
@@ -1,6 +1,8 @@
 module.exports = function(client, callback) {
-	dbAliad.query('select * from City', function(err, rows, fields) {
+
+	dbAlias.query('select * from City', function(err, rows, fields) {
 		client.context.data = { rows:rows, fields:fields };
 		callback();
 	});
+
 }