+++ examples/copyContentToProjectFolder/applications/example/app/examples/security/getUsers.json/get.js
@@ -1,6 +1,8 @@
 module.exports = function(client, callback) {
+
 	impress.security.db.users.find({}).toArray(function(err, nodes) {
 		client.context.data = nodes;
 		callback();
 	});
+
 }