+++ examples/copyContentToProjectFolder/applications/example/app/examples/simple/ajaxTest.ajax/get.js
@@ -1,6 +1,8 @@
 module.exports = function(client, callback) {
+
 	client.context.data = {
 		parameterName: client.query.parameterName,
 	};
 	callback();
+
 }