+++ examples/copyContentToProjectFolder/applications/example/app/examples/simple/uploadFile.json/post.js
@@ -1,4 +1,6 @@
 ﻿module.exports = function(client, callback) {
+
 	client.context.data = client.files;
 	callback();
+
 }