+++ examples/copyContentToProjectFolder/applications/example/app/examples/tools/forkWorker.json/worker.js
@@ -1,6 +1,7 @@
 module.exports = function(client, callback) {
 
     console.log('Message from forked worker');
+    console.dir(client);
 	callback();
 
 }