+++ examples/copyContentToProjectFolder/applications/example/app/examples/tools/serverHealth.json/get.js
@@ -1,7 +1,6 @@
 module.exports = function(client, callback) {
 
 	client.application.leak = client.application.leak || [];
-	// console.log(client.application.leak.length);
 	var n = client.application.leak.length ? client.application.leak.length : 0;
 	client.application.leak.push(new Array(1024*(1024/n+1)).join("Hello node.js, Hello Impress Application Server "));